p. 177 add test with second auction; make startBiddingIn accept multiple auctions and pass their IDs to Main
Note that on line 35 of ApplicationRunner, in the call to driver.showsSniperStatus, the first argument is empty string ("") instead of auction.getItemId() as shown in the book. There is no code to show the ID upon initially joining yet, so the code as in the book fails all tests and produces a different error message than expected.