p. 181 add tests to make sure correct row in sniper table model is updated for a given snapshot; item ID  should appear when joining
The updatesCorrectRowForSniper test is not shown in the book and the version in https://github.com/sf105/goos-code/blob/master/src/auctionsniper/ui/SnipersTableModel.java is in a later stage of development, so I had to take a guess at what the original test may have looked like.

Also, at this point, the item ID will be displayed when joining, so ApplicationRunner has to be updated. This is not mentioned in the book, but the code was shown on p. 177, at which time it caused a premature failure with a different error than shown in the book because the table model was still using an initial default at that point.